+++ examples/calc.ml
@@ -21,8 +21,7 @@
       parse "123*123 - 23*234";;
     v} *)
 
-module P = Reparse.Parser
-open P.Infix
+open Reparse.Parser
 
 type expr =
   | Int of int
@@ -31,41 +30,41 @@ type expr =
   | Mult of expr * expr
   | Div of expr * expr
 
-let skip_spaces = P.skip P.space
+let skip_spaces = skip space >>| fun _ -> ()
 
-let binop : 'a P.t -> char -> 'b P.t -> ('a -> 'b -> 'c) -> 'c P.t =
+let binop : 'a t -> char -> 'b t -> ('a -> 'b -> 'c) -> 'c t =
  fun exp1 op exp2 f ->
-  P.map3 (fun e1 _ e2 -> f e1 e2) exp1 (skip_spaces *> P.char op <* skip_spaces) exp2
+  map3 exp1 (skip_spaces *> char op <* skip_spaces) exp2 ~f:(fun e1 _ e2 -> f e1 e2)
 ;;
 
-let integer : expr P.t =
-  let+ d = P.digits in
+let integer : expr t =
+  let+ d = digits in
   Int (int_of_string d)
 ;;
 
-let factor : expr P.t -> expr P.t =
+let factor : expr t -> expr t =
  fun expr ->
-  P.any
-    [ P.char '(' *> skip_spaces *> expr <* skip_spaces <* P.char ')'
+  any
+    [ char '(' *> skip_spaces *> expr <* skip_spaces <* char ')'
     ; skip_spaces *> integer <* skip_spaces
     ]
 ;;
 
-let term : expr P.t -> expr P.t =
+let term : expr t -> expr t =
  fun factor ->
-  P.recur (fun term ->
+  recur (fun term ->
       let mult = binop factor '*' term (fun e1 e2 -> Mult (e1, e2)) in
       let div = binop factor '/' term (fun e1 e2 -> Div (e1, e2)) in
       mult <|> div <|> factor)
 ;;
 
-let expr : expr P.t =
-  P.recur (fun expr ->
+let expr : expr t =
+  recur (fun expr ->
       let factor = factor expr in
       let term = term factor in
       let add = binop term '+' expr (fun e1 e2 -> Add (e1, e2)) in
       let sub = binop term '-' expr (fun e1 e2 -> Sub (e1, e2)) in
-      P.any [ add; sub; term ])
+      any [ add; sub; term ])
 ;;
 
 let rec eval : expr -> int = function
@@ -78,14 +77,14 @@ let rec eval : expr -> int = function
 
 (* Test AST *)
 let r =
-  let actual = P.parse_string expr "1*2-4+3" in
+  let actual = parse_string expr "1*2-4+3" in
   let expected = Sub (Mult (Int 1, Int 2), Add (Int 4, Int 3)) in
   Bool.equal (expected = actual) true
 ;;
 
 (* Run and test the evaluator. *)
 let exp_result =
-  let v = eval (P.parse_string expr "12+1*10") in
+  let v = eval (parse_string expr "12+1*10") in
   Int.equal 22 v
 ;;